requirements.txt
# ========================================
# QORICASH TRADING V2 - DEPENDENCIAS
# ========================================

# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0

# Security
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# WebSocket (Real-time)
Flask-SocketIO==5.3.5
python-socketio==5.10.0
simple-websocket==1.0.0

# File Upload
cloudinary==1.41.0

# Data Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Excel Export
openpyxl==3.1.2

# Timezone
pytz==2024.1

# Production Server
gunicorn==21.2.0
eventlet==0.35.2

# Email
Flask-Mail==0.9.1

# Testing (opcional)
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
